==> 0_data_load.ipynb <==
import json
import pandas as pd

with open("data/raw/datasenado_json_completo.json", "r", encoding='utf-8-sig') as f:
    data = json.load(f)

clean_data= []

for bloco in data['perguntas']:
    for perguntas in bloco['perguntas']:
        for ano in perguntas['anos']:
            for cat in ano['categorias']:
                row = {}
                row['bloco'] = bloco['bloco']
                row['question']= perguntas['titulo']
                row['year']= ano['ano']
                row['answer'] = cat['cat']
                row['percentage'] = cat['perc']
                clean_data.append(row)

clean_data_df = pd.DataFrame(clean_data)

clean_data_df.head(5)

# save to csv
clean_data_df.to_csv('data/clean/clean_df.csv', index=False)
