==> 1_data_transformation.ipynb <==
## 0. Libraries 

import pandas as pd
import matplotlib
import numpy as np  
from googletrans import Translator
from textblob import TextBlob
from deep_translator import GoogleTranslator
import time
from textblob import TextBlob

## 1. Loading Data

# Load the clean data
df = pd.read_csv('data/clean/clean_df.csv')

## 2. Data Understandig 

- Numbers of interviews 

- 2005 - 815
- 2007 - 797
- 2009 - 827 
- 2011 - 1352
- 2013 - 1248
- 2015 - 1102
- 2017 - 1116
- 2019 - 2400
- 2021 - 3000
- 2023 - 21787



df.info()  

df.duplicated().sum()  # Check for duplicates

df.head()   # Check the first few rows of the data

## 3. Data Transformation 

# Define the function for translation
def translate_text(text, source_lang="pt", target_lang="en"):
    """Translates text using GoogleTranslator from deep-translator."""
    try:
        if pd.notna(text):  
            return GoogleTranslator(source=source_lang, target=target_lang).translate(text)
        return text
    except Exception as e:
        print(f"Translation failed for '{text}': {e}")
        return text  

# Ensure profile_df exists
if "df1" not in locals():
    df1 = df.copy()  

# Apply translation to relevant columns
df1["category_translated"] = df["bloco"].apply(translate_text)
df1["question_translated"] = df["question"].apply(translate_text)
df1["answer_translated"] = df["answer"].apply(translate_text)


# Dictionary with the number of interviews per year
interviews_per_year = {
    2005: 815,
    2007: 797,
    2009: 827,
    2011: 1352,
    2013: 1248,
    2015: 1102,
    2017: 1116,
    2019: 2400,
    2021: 3000,
    2023: 21787
}

# Add a new column calculating the total number of answers for each row
df1["total_answers"] = df1.apply(lambda row: int((row["percentage"] / 100) * interviews_per_year.get(row["year"], 1)), axis=1)

# Save the translated DataFrame
df1.to_csv("data/clean/translated_df.csv", index=False)

df2 = df1.copy()
df2.drop(columns=["bloco", "question", "answer"], inplace=True)
df2.rename(columns={"category_translated": "category", "question_translated": "question", "answer_translated": "answer"}, inplace=True)
df2[['year', 'category', 'question', 'answer', 'percentage']]
df2["year"] = df2["year"].astype(str)

# Create a new DataFrame with selected columns from df2
df3 = df2[['year', 'category', 'question', 'answer', 'percentage', 'total_answers']]
df3

# Save the final DataFrame
df3.to_csv("data/clean/final_translated_df.csv", index=False)
