==> 3_new_data.ipynb <==
import json
import pandas as pd

with open("data/raw/datasenado_json_completo.json", "r", encoding='utf-8-sig') as f:
    data = json.load(f)

question_1 = 'Você já sofreu algum tipo de violência doméstica ou familiar provocada por um homem?'

clean_data = []

for bloco in data['perguntas']:
    if bloco['bloco'] == 'A violência vivida / testemunhada':
        for pergunta in bloco['perguntas']:
            if pergunta['titulo'] == question_1:
                for ano in pergunta['anos']:
                    for cruzamento in ano['cruzamentos']:
                        for categorias in cruzamento['categorias']:
                            for subcategoria in categorias['subcategorias']:
                                row = {}
                                row['ano'] = ano['ano']
                                row['var_cruz'] = cruzamento['var_cruz']
                                row['categoria'] = categorias['categoria']
                                row['subcategoria'] = subcategoria['cat']
                                row['percentual'] = subcategoria['perc'] 
                                clean_data.append(row)

clean_data_df = pd.DataFrame(clean_data)

clean_data_df.head()

# Save data to csv
clean_data_df.to_csv('data/clean/question_1.csv', index=False)
