==> 4_data_question_transformation.ipynb <==
## 0. Libraries 

import pandas as pd
import matplotlib
import numpy as np  
import matplotlib.pyplot as plt
import seaborn as sns
from googletrans import Translator
from textblob import TextBlob
from deep_translator import GoogleTranslator
import time
from textblob import TextBlob

## 1. Load Data 

# Load the data
df = pd.read_csv('data/clean/question_1.csv')

## 2. Data Transformation 

# Define the function for translation
def translate_text(text, source_lang="pt", target_lang="en"):
    """Translates text using GoogleTranslator from deep-translator."""
    try:
        if pd.notna(text):  # Check if text is not NaN
            return GoogleTranslator(source=source_lang, target=target_lang).translate(text)
        return text
    except Exception as e:
        print(f"Translation failed for '{text}': {e}")
        return text  # Keep the original text if translation fails

# Ensure profile_df exists
if "df1" not in locals():
    df1 = df.copy()  # Copy the original DataFrame

# Apply translation to relevant columns
df1["category"] = df["var_cruz"].apply(translate_text)
df1["answer"] = df["categoria"].apply(translate_text)
df1["subcategory"] = df["subcategoria"].apply(translate_text)

# Drop columns that are not needed
df1.drop(columns=["var_cruz", "categoria", "subcategoria"], inplace=True)

# Rename columns
df1.rename(columns={"ano": "year", "percentual":"percentage"} , inplace=True)

# Reorder columns
df1 = df1[["year", "category", "subcategory", "answer", "percentage"]]

# Save the translated DataFrame
df1.to_csv('data/clean/question_1_translated.csv', index=False)
